requirements.txt
discord.py
pycrypto
